재고관리 입출고 일> 일시로 변경
- 재고관리 일자 > 일시로 변경
- 재고 조정 사유 이넘 분리
- 가격 데이터타입 Long> BigDecimal로 변경
- columnDefinition comment > @Comment로 변경
- 재고 조정 - 모달로 변경

+++ src/main/java/com/boot/ict05_final_admin/domain/inventory/dto/InventoryInWriteDTO.java
@@ -18,10 +18,10 @@ public class InventoryInWriteDTO {
     private BigDecimal quantity;
 
     /** 입고 단가 (본사 매입가) */
-    private Long unitPrice;
+    private BigDecimal unitPrice;
 
     /** 출고 단가 (가맹점 공급가) */
-    private Long sellingPrice;
+    private BigDecimal sellingPrice;
 
     /** 입고일 */
     private LocalDateTime inDate;
